+++ addon/components/power-select-multiple/trigger.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
 import layout from '../../templates/components/power-select-multiple/trigger';
 import { emberPowerSelectBuildSelection as buildNewSelection } from '../../helpers/ember-power-select-build-selection';
+import updateInput from '../../utils/update-input-value';
 
 const { computed, get, isBlank, run } = Ember;
 const { htmlSafe } = Ember.String;
@@ -10,10 +11,18 @@ export default Ember.Component.extend({
   layout,
 
   // Lifecycle hooks
+  didInsertElement() {
+    this._super(...arguments);
+    this.input = document.querySelector(`.${this.elementId}-input`);
+  },
+
   didUpdateAttrs({ oldAttrs, newAttrs }) {
     this._super(...arguments);
     if (oldAttrs.select.isOpen && !newAttrs.select.isOpen) {
       this.handleClose();
+    }
+    if (newAttrs.searchText !== oldAttrs.searchText) {
+      run.scheduleOnce('afterRender', this, this.updateInput, oldAttrs.searchText, newAttrs.searchText);
     }
   },
 
@@ -35,7 +44,7 @@ export default Ember.Component.extend({
   // Actions
   actions: {
     search(term, e) {
-      const { search, open } = this.get('select.actions');
+      let { search, open } = this.get('select.actions');
       search(term, e);
       open(e);
     },
@@ -71,5 +80,9 @@ export default Ember.Component.extend({
   // Methods
   handleClose() {
     run.scheduleOnce('actions', null, this.get('select.actions.search'), '');
+  },
+
+  updateInput(oldText, newText) {
+    updateInput(this.input, oldText, newText);
   }
 });